requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
tensorflow==2.21.0
tf-keras==2.21.0
pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9